// scripts/src/kind/types.ts
export interface CompanyInfo {
  companyName: string; // 회사명
  stockCode: number; // 종목코드
  industry: string; // 업종
  mainProducts: string; // 주요제품
  listingDate: string; // 상장일
  fiscalMonth: string; // 결산월
  CEOName: string; // 대표자명
  website: string; // 홈페이지
  location: string; // 지역
}